Consolidate honeypot, security & throttle inside filters
