+++ application/Filters/Honeypot.php
@@ -4,7 +4,8 @@ use CodeIgniter\Filters\FilterInterface;
 use CodeIgniter\HTTP\RequestInterface;
 use CodeIgniter\HTTP\ResponseInterface;
 use Config\Services;
-use CodeIgniter\Honeypot\Exceptions\HoneypotException;
+use CodeIgniter\Filters\Exceptions\HoneypotException;
+use CodeIgniter\Filters\Honeypot;
 
 class Honeypot implements FilterInterface 
 {
