+++ system/Filters/Exceptions/HoneypotException.php
@@ -1,4 +1,4 @@
-<?php namespace CodeIgniter\Honeypot\Exceptions;
+<?php namespace CodeIgniter\Filters\Exceptions;
 
 use CodeIgniter\Exceptions\ConfigException;
 use CodeIgniter\Exceptions\ExceptionInterface;
